auto_scripts/func-script/requirements.txt
# Core testing framework
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.3.1
pytest-rerunfailures==12.0

# Selenium WebDriver and related packages
selenium==4.15.2
webdriver-manager==4.0.1

# Configuration and data handling
PyYAML==6.0.1
jsonschema==4.19.2

# HTTP requests for API testing
requests==2.31.0
urllib3==2.0.7

# Reporting and logging
allure-pytest==2.13.2
logging==0.4.9.6

# Email functionality
secure-smtplib==0.1.1

# Data manipulation and utilities
pandas==2.1.3
numpy==1.25.2
openpyxl==3.1.2

# Image processing for screenshots
Pillow==10.1.0

# Date and time utilities
python-dateutil==2.8.2

# Environment variable management
python-dotenv==1.0.0

# JSON manipulation
jsonpath-ng==1.6.0

# Faker for test data generation
Faker==19.12.0

# Retry mechanisms
retrying==1.3.4

# Performance testing (optional)
locust==2.17.0

# Code quality tools (development)
flake8==6.1.0
black==23.10.1
pytest-cov==4.1.0

# Additional utilities
colorlog==6.7.0
tqdm==4.66.1
click==8.1.7